+++ src/utils/dlb.c
@@ -34,7 +34,7 @@ static void print_usage( const char * program ) {
 static void print_help( void ) {
     fprintf( stdout, "DLB - Dynamic Load Balancing, version %s.\n", VERSION );
     fprintf( stdout, "\n" );
-    setenv("LB_POLICY", "no", 0);
+    /* setenv("LB_POLICY", "no", 0); */
     DLB_Init();
     DLB_PrintVariables();
     DLB_Finalize();